refactor(adr-061): give the namespace-glob matcher its own module
`suppression.py` sat seven lines under the 2000-line hard cap. The one
self-contained cluster inside it is the namespace-glob matcher: compiling a
`foo::bar::*` pattern, walking a symbol against it without backtracking, and
the character-level helpers under both. It is a general string-matching
primitive that knows nothing about suppressions, findings or policy, and
`suppression.py` was its only caller.

Moved whole to `policy/namespace_glob.py` -- 1993 -> 1355 lines, the largest
single reduction in this PR. The pieces move together because they are one
mechanism: splitting `_translate_namespace_glob` from `_SegmentGlobMatcher`
from `_match_run` would leave a caller reaching across modules for every
step. Every name is re-exported under its original spelling, which
`tests/test_suppression_edge_cases.py` needs -- it imports
`_SEGMENT_RE_WRAPPER` and `_fnmatch_segment_regex` from `suppression`
directly.

Root `AGENTS.md`'s "Primitive-level property tests" section asks for a
standalone property-test class whenever a reusable primitive gets its own
home, so `tests/unit/policy/test_namespace_glob_properties.py` states the
contract as invariants over an exhaustively enumerated small domain (~2000
pattern/name pairs, under a second).

The central one is differential, and its reach was measured rather than
assumed. The module carries two spellings of one semantics -- a regex
translation and a non-backtracking walk written to replace it for
performance while "exactly reproducing" its behaviour -- and every caller
uses only the walk, so the two could drift with nothing in production
noticing. Seeding drift three ways showed what the sweep does and does not
catch: a change to the regex half's own body fails it, a change to the walk's
own body fails it, and a change to the shared `_fnmatch_segment_regex` does
*not*, because both halves call it. The module docstring says so plainly
rather than claiming "two independent implementations", and points at the
existing direct assertions on those shared helpers.

Verified: `check_architecture.py` 0 findings, `ruff` and `mypy abicheck/`
clean (575 files), 1233 suppression/glob tests pass.
